+++ workspaces/azure-devops/plugins/azure-devops-backend/src/api/AzureDevOpsApi.ts
@@ -215,6 +215,9 @@ export class AzureDevOpsApi {
     );
   }
 
+  /**
+   * @deprecated This method has no usages and will be removed in a future release
+   */
   public async getRepoBuilds(
     projectName: string,
     repoName: string,
@@ -458,6 +461,9 @@ export class AzureDevOpsApi {
     }));
   }
 
+  /**
+   * @deprecated This method has no usages and will be removed in a future release
+   */
   public async getBuildDefinitions(
     projectName: string,
     definitionName: string,