+++ workspaces/azure-devops/plugins/azure-devops/src/hooks/useReadme.ts
@@ -20,7 +20,7 @@ import { Entity, stringifyEntityRef } from '@backstage/catalog-model';
 import { azureDevOpsApiRef } from '../api';
 import { useApi } from '@backstage/core-plugin-api';
 import useAsync from 'react-use/esm/useAsync';
-import { getAnnotationValuesFromEntity } from '../utils';
+import { getAnnotationValuesFromEntity } from '@backstage-community/plugin-azure-devops-common';
 
 export function useReadme(entity: Entity): {
   item?: Readme;
@@ -32,6 +32,13 @@ export function useReadme(entity: Entity): {
   const { value, loading, error } = useAsync(() => {
     const { project, repo, host, org, readmePath } =
       getAnnotationValuesFromEntity(entity);
+
+    if (readmePath?.startsWith('.')) {
+      throw new Error(
+        `The "dev.azure.com/readme-path" annotation does not support relative paths, please correct this annotation. The value provided was: "${readmePath}"`,
+      );
+    }
+
     const entityRef = stringifyEntityRef(entity);
     return api.getReadme({
       project,