+++ workspaces/rbac/plugins/rbac-backend/src/service/policies-rest-api.test.ts
@@ -25,7 +25,6 @@ import request from 'supertest';
 import {
   PermissionAction,
   PermissionInfo,
-  PluginPermissionMetaData,
   policyEntityCreatePermission,
   policyEntityDeletePermission,
   policyEntityReadPermission,
@@ -55,13 +54,17 @@ import {
   mockedAuthorizeConditional,
   mockHttpAuth,
   mockLoggerService,
-  mockPermissionEvaluator,
   mockUserInfoService,
+  mockPermissionEvaluator,
   pluginMetadataCollectorMock,
   providerMock,
   roleMetadataStorageMock,
   mockedAuthorize,
+  mockPermissionRegistry,
+  permissionDependentPluginStoreMock,
+  extendablePluginIdProviderMock,
 } from '../../__fixtures__/mock-utils';
+import { ExtendablePluginIdProvider } from './extendable-id-provider';
 
 jest.setTimeout(60000);
 
@@ -116,7 +119,7 @@ const expectedConditions: RoleConditionalPolicyDecision<PermissionAction>[] = [
 
 const modifiedBy = 'user:default/some-admin';
 
-describe('REST policies api', () => {
+describe('REST policies API', () => {
   let app: express.Express;
   let config = mockServices.rootConfig({
     data: {
@@ -258,16 +261,19 @@ describe('REST policies api', () => {
         mockAuthService,
       ),
       userInfo: mockUserInfoService,
+      permissionsRegistry: mockPermissionRegistry,
+      auditor: mockAuditorService,
+      permissions: mockPermissionEvaluator,
     };
 
     server = new PoliciesServer(
-      mockPermissionEvaluator,
       options,
       enforcerDelegateMock as EnforcerDelegate,
       conditionalStorageMock,
       pluginMetadataCollectorMock as PluginPermissionMetadataCollector,
       roleMetadataStorageMock,
-      mockAuditorService,
+      permissionDependentPluginStoreMock,
+      extendablePluginIdProviderMock as ExtendablePluginIdProvider,
     );
     const router = await server.serve();
     app = express().use(router);
@@ -3839,16 +3845,19 @@ describe('REST policies api', () => {
           mockAuthService,
         ),
         userInfo: mockUserInfoService,
+        permissionsRegistry: mockPermissionRegistry,
+        auditor: mockAuditorService,
+        permissions: mockPermissionEvaluator,
       };
 
       server = new PoliciesServer(
-        mockPermissionEvaluator,
         options,
         enforcerDelegateMock as EnforcerDelegate,
         conditionalStorageMock,
         pluginMetadataCollectorMock as PluginPermissionMetadataCollector,
         roleMetadataStorageMock,
-        mockAuditorService,
+        permissionDependentPluginStoreMock,
+        extendablePluginIdProviderMock as ExtendablePluginIdProvider,
         [providerMock],
       );
       const router = await server.serve();
@@ -3905,112 +3914,6 @@ describe('REST policies api', () => {
       expect(result.body.error).toEqual({
         message: 'The RBAC provider test was not found',
         name: 'NotFoundError',
-      });
-    });
-  });
-
-  describe('list plugin permissions and condition rules', () => {
-    it('should return list plugins permission', async () => {
-      const pluginMetadata: PluginPermissionMetaData[] = [
-        {
-          pluginId: 'permissions',
-          policies: [
-            {
-              name: 'catalog.entity.read',
-              resourceType: 'policy-entity',
-              policy: 'read',
-            },
-          ],
-        },
-      ];
-      pluginMetadataCollectorMock.getPluginPolicies = jest
-        .fn()
-        .mockImplementation(async () => {
-          return pluginMetadata;
-        });
-      const result = await request(app).get('/plugins/policies').send();
-      expect(result.statusCode).toEqual(200);
-      expect(result.body).toEqual(pluginMetadata);
-    });
-
-    it('should return a status of Unauthorized for /plugins/policies', async () => {
-      mockedAuthorizeConditional.mockImplementationOnce(async () => [
-        { result: AuthorizeResult.DENY },
-      ]);
-      const result = await request(app).get('/plugins/policies').send();
-
-      expect(mockedAuthorizeConditional).toHaveBeenCalledWith(
-        [
-          {
-            permission: policyEntityReadPermission,
-          },
-        ],
-        {
-          credentials: credentials,
-        },
-      );
-      expect(result.statusCode).toBe(403);
-      expect(result.body.error).toEqual({
-        name: 'NotAllowedError',
-        message: '',
-      });
-    });
-
-    it('should return list plugins condition rules', async () => {
-      const rules: PluginMetadataResponseSerializedRule[] = [
-        {
-          pluginId: 'catalog',
-          rules: [
-            {
-              description: 'Allow entities with the specified label',
-              name: 'HAS_LABEL',
-              paramsSchema: {
-                $schema: 'http://json-schema.org/draft-07/schema#',
-                additionalProperties: false,
-                properties: {
-                  label: {
-                    description: 'Name of the label to match on',
-                    type: 'string',
-                  },
-                },
-                required: ['label'],
-                type: 'object',
-              },
-              resourceType: 'catalog-entity',
-            },
-          ],
-        },
-      ];
-      pluginMetadataCollectorMock.getPluginConditionRules = jest
-        .fn()
-        .mockImplementation(async () => {
-          return rules;
-        });
-      const result = await request(app).get('/plugins/condition-rules').send();
-      expect(result.statusCode).toEqual(200);
-      expect(result.body).toEqual(rules);
-    });
-
-    it('should return a status of Unauthorized for /plugins/condition-rules', async () => {
-      mockedAuthorizeConditional.mockImplementationOnce(async () => [
-        { result: AuthorizeResult.DENY },
-      ]);
-      const result = await request(app).get('/plugins/condition-rules').send();
-
-      expect(mockedAuthorizeConditional).toHaveBeenCalledWith(
-        [
-          {
-            permission: policyEntityReadPermission,
-          },
-        ],
-        {
-          credentials: credentials,
-        },
-      );
-      expect(result.statusCode).toBe(403);
-      expect(result.body.error).toEqual({
-        name: 'NotAllowedError',
-        message: '',
       });
     });
   });