+++ workspaces/adr/plugins/adr/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @backstage-community/plugin-adr
 
+## 0.11.0
+
+### Minor Changes
+
+- fc4fa2c: Backstage version bump to v1.39.0
+
+### Patch Changes
+
+- Updated dependencies [fc4fa2c]
+  - @backstage-community/plugin-adr-common@0.8.0
+
+## 0.10.0
+
+### Minor Changes
+
+- e22d803: Backstage version bump to v1.38.1
+
+### Patch Changes
+
+- Updated dependencies [e22d803]
+  - @backstage-community/plugin-adr-common@0.7.0
+
 ## 0.9.2
 
 ### Patch Changes