+++ workspaces/analytics/plugins/analytics-module-newrelic-browser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-analytics-module-newrelic-browser
 
+## 0.6.0
+
+### Minor Changes
+
+- 8b665f9: Backstage version bump to v1.38.1
+
 ## 0.5.1
 
 ### Patch Changes