+++ workspaces/announcements/plugins/announcements-backend/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @backstage-community/plugin-announcements-backend
 
+## 0.5.0
+
+### Minor Changes
+
+- 7e38fa0: Backstage version bump to v1.39.1
+
+### Patch Changes
+
+- Updated dependencies [7e38fa0]
+  - @backstage-community/plugin-search-backend-module-announcements@0.4.0
+  - @backstage-community/plugin-announcements-common@0.5.0
+
+## 0.4.0
+
+### Minor Changes
+
+- 9c94358: Added support for submitting announcements on behalf of a team the user is a member of.
+  This field is optional. if not specified, the announcement will be submitted using the current publisher user.
+
+### Patch Changes
+
+- Updated dependencies [9c94358]
+  - @backstage-community/plugin-announcements-common@0.4.0
+  - @backstage-community/plugin-search-backend-module-announcements@0.3.4
+
 ## 0.3.3
 
 ### Patch Changes