+++ workspaces/graphql-voyager/plugins/graphql-voyager/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-graphql-voyager
 
+## 0.7.0
+
+### Minor Changes
+
+- 10f80d4: Backstage version bump to v1.39.0
+
 ## 0.6.0
 
 ### Minor Changes