+++ workspaces/lighthouse/plugins/lighthouse/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage-community/plugin-lighthouse
 
+## 0.10.0
+
+### Minor Changes
+
+- c2d01f5: Backstage version bump to v1.39.0
+
+### Patch Changes
+
+- Updated dependencies [c2d01f5]
+  - @backstage-community/plugin-lighthouse-common@0.7.0
+
 ## 0.9.0
 
 ### Minor Changes