+++ workspaces/microsoft-calendar/plugins/microsoft-calendar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-microsoft-calendar
 
+## 0.7.0
+
+### Minor Changes
+
+- 11f97a9: Backstage version bump to v1.39.0
+
 ## 0.6.0
 
 ### Minor Changes