+++ workspaces/multi-source-security-viewer/plugins/multi-source-security-viewer/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @backstage-community/plugin-multi-source-security-viewer
 
+## 0.5.0
+
+### Minor Changes
+
+- 2284286: Bump version to 1.39.1
+
+### Patch Changes
+
+- Updated dependencies [2284286]
+  - @backstage-community/plugin-multi-source-security-viewer-common@0.3.0
+
+## 0.4.3
+
+### Patch Changes
+
+- bec7d2b: Updated dependency `@backstage-community/plugin-jenkins` to `^0.19.0`.
+
+## 0.4.2
+
+### Patch Changes
+
+- ca224e7: Updated dependency `@backstage-community/plugin-azure-devops-common` to `^0.10.0`.
+
 ## 0.4.1
 
 ### Patch Changes