+++ workspaces/scaffolder-backend-module-servicenow/plugins/scaffolder-backend-module-servicenow/CHANGELOG.md
@@ -1,5 +1,17 @@
 ### Dependencies
 
+## 2.6.3
+
+### Patch Changes
+
+- fd39ec9: Updated dependency `@hey-api/openapi-ts` to `0.72.2`.
+
+## 2.6.2
+
+### Patch Changes
+
+- 31b91ce: Updated dependency `@hey-api/openapi-ts` to `0.67.3`.
+
 ## 2.6.1
 
 ### Patch Changes