+++ workspaces/sonarqube/plugins/sonarqube-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage-community/plugin-sonarqube-react
 
+## 0.7.0
+
+### Minor Changes
+
+- f3284e7: Backstage version bump to v1.39.0
+
+## 0.6.0
+
+### Minor Changes
+
+- 8be1894: Upgraded to Backstage release 1.38
+
 ## 0.5.0
 
 ### Minor Changes